+++ StoredProcedures/StoredProcedures_Session.sql
@@ -5,10 +5,10 @@ create procedure SP_Se_Register
     @email VARCHAR(1000),
     @birthdate DATETIME,
     @hash varchar(1000),
-    @gender int,
-    @languageId int,
-    @locationId int,
-    @roleId int
+    @gender bigint,
+    @languageId bigint,
+    @locationId bigint,
+    @roleId bigint
 AS
     INSERT INTO Person (username, firstname, lastname, email, birthdate, hash, gender, LanguageID, LocationID, RoleID) VALUES
     (@username, @firstname, @lastname, @email, Convert(datetime, @birthdate), @hash, @gender, convert(int, @languageId),
